requirements.txt
django
langchain-community
pinecone-client
langchain_openai
langchain
whitenoise